Refactor package dao to repositories
