+++ PURL/src/main/java/com/hida/repositories/CitationDao.java
@@ -1,4 +1,4 @@
-package com.hida.dao;
+package com.hida.repositories;
 
 import com.hida.model.Citation;
