+++ PURL/src/main/java/com/hida/repositories/CitationDaoImpl.java
@@ -1,4 +1,4 @@
-package com.hida.dao;
+package com.hida.repositories;
 
 import com.hida.model.Citation;
 import org.springframework.stereotype.Repository;
